+++ data_science_pipeline/cli/peerscout_build_reviewing_editor_profiles.py
@@ -1,0 +1,16 @@
+import logging
+from data_science_pipeline.utils.notebook import run_notebook
+
+logging.basicConfig(level=logging.INFO)
+LOGGER = logging.getLogger(__name__)
+
+
+def main():
+    run_notebook(
+        notebook_filename='peerscout/peerscout-build-reviewing-editor-profiles.ipynb'
+    )
+
+
+if __name__ == '__main__':
+    logging.basicConfig(level=logging.INFO)
+    main()
